requirements.txt
boto3==1.24.18
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
werkzeug
Flask-Login==0.6.1
flask-wtf==1.0.1
wtforms
email-validator